backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.20
alembic==1.12.0

# Document Processing
pytesseract==0.3.10
Pillow==10.0.1
PyMuPDF==1.23.0
python-docx==0.8.11
openpyxl==3.1.2
python-pptx==0.6.21
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'

# Configuration and validation
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Utilities
pandas==2.1.0
numpy==1.24.3
python-dateutil==2.8.2
python-slugify==8.0.1
tqdm==4.66.1

# Security
cryptography==41.0.7
passlib==1.7.4

# Optional AI/ML (comment out if causing issues)
# scikit-learn==1.3.2
# sentence-transformers==2.2.2
# transformers==4.33.3
# torch==2.0.1
# faiss-cpu==1.7.4
# whoosh==2.7.4

# Development
pytest==7.4.2
black==23.7.0
flake8==6.1.0
mypy==1.5.1
